Conflicts:
	Studies/FierzOctetAnalyzer.cc

+++ Analysis/G4toPMT.hh
@@ -59,7 +59,7 @@ public:
 	double costheta;			//< primary event cos pitch angle
 	double ePrim;				//< primary event energy
 	unsigned int nToSim;		//< total number of events to simulate (set to 0 for random energy selection)
-	unsigned int nSimmed;		//< number of events simulated since scan start
+	double nSimmed;				//< number of events simulated since scan start
 	double mwpcThresh[2];		//< MWPC trigger threshold on each side
 	
 protected:
